refactor: split GraphQLExplorer into GraphQLEndpointForm and GraphiQLInterface
